# api-gateway-user/docs/swagger.yaml
basePath: /
definitions:
...
  genproto.BookCreate:
    properties:
      Description:
        type: string
      PictureUrls:
        items:
          type: string
        type: array
      Price:
        type: number
      Title:
        type: string
    type: object
  genproto.BookGet:
    properties:
      CreatedAt:
        type: string
      Description:
        type: string
      Id:
        type: string
      PictureUrls:
        items:
          type: string
        type: array
      Price:
        type: number
      Title:
        type: string
    type: object
  genproto.BookList:
    properties:
      Books:
        items:
          $ref: '#/definitions/genproto.BookGet'
        type: array
      Pagination:
        $ref: '#/definitions/genproto.Pagination'
      TotalCount:
        type: integer
    type: object
  genproto.BookPicture:
    properties:
      BookId:
        type: string
      PictureUrl:
        type: string
    type: object
  genproto.BookUpt:
    properties:
      Description:
        type: string
      Price:
        type: number
      Title:
        type: string
    type: object
...
  genproto.MaterialCreateReq:
    properties:
      description:
        type: string
      lesson_id:
        type: string
      resource_url:
        type: string
      title:
        type: string
    type: object
  genproto.MaterialGetRes:
    properties:
      description:
        type: string
      id:
        type: string
      lesson_id:
        type: string
      resource_url:
        type: string
      title:
        type: string
    type: object
  genproto.MaterialListRes:
    properties:
      materials:
        items:
          $ref: '#/definitions/genproto.MaterialGetRes'
        type: array
    type: object
  genproto.MaterialUpdate:
    properties:
      description:
        type: string
      resource_url:
        type: string
      title:
        type: string
    type: object
...
  genproto.Pagination:
    properties:
      limit:
        type: integer
      offset:
        type: integer
    type: object
...
info:
  contact: {}
  description: API for Instant Delivery resources
  title: Finance Tracker API Documentation
  version: "1.0"
paths:
  /books/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a book by its ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Book details
          schema:
            $ref: '#/definitions/genproto.BookGet'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Book
      tags:
      - Books
  /books/create:
    post:
      consumes:
      - application/json
      description: Create a new book
      parameters:
      - description: Book data
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/genproto.BookCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Book created successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Book
      tags:
      - Books
  /books/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a book by its ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Book deleted successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Book
      tags:
      - Books
  /books/list:
    get:
      consumes:
      - application/json
      description: Retrieve a list of books
      parameters:
      - description: Title
        in: query
        name: title
        type: string
      - description: Price From
        in: query
        name: price_from
        type: number
      - description: Price To
        in: query
        name: price_to
        type: number
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of books
          schema:
            $ref: '#/definitions/genproto.BookList'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List Books
      tags:
      - Books
  /books/picture/add:
    post:
      consumes:
      - application/json
      description: Add a picture to a book
      parameters:
      - description: Book picture data
        in: body
        name: bookPicture
        required: true
        schema:
          $ref: '#/definitions/genproto.BookPicture'
      produces:
      - application/json
      responses:
        "200":
          description: Picture added successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Add Picture
      tags:
      - Books
  /books/picture/delete:
    delete:
      consumes:
      - application/json
      description: Delete a picture from a book
      parameters:
      - description: Book picture data
        in: body
        name: bookPicture
        required: true
        schema:
          $ref: '#/definitions/genproto.BookPicture'
      produces:
      - application/json
      responses:
        "200":
          description: Picture deleted successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Picture
      tags:
      - Books
  /books/update/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing book
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      - description: Book data
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/genproto.BookUpt'
      produces:
      - application/json
      responses:
        "200":
          description: Book updated successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Book
      tags:
      - Books
...
  /materials/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a material by its ID
      parameters:
      - description: Material ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Material deleted successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Material not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Material
      tags:
      - Material
    get:
      consumes:
      - application/json
      description: Retrieve a material by its ID
      parameters:
      - description: Material ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Material details
          schema:
            $ref: '#/definitions/genproto.MaterialGetRes'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Material not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Material
      tags:
      - Material
    put:
      consumes:
      - application/json
      description: Update an existing material
      parameters:
      - description: Material ID
        in: path
        name: id
        required: true
        type: string
      - description: Material data
        in: body
        name: material
        required: true
        schema:
          $ref: '#/definitions/genproto.MaterialUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Material updated successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Material
      tags:
      - Material
  /materials/create:
    post:
      consumes:
      - application/json
      description: Create a new material
      parameters:
      - description: Material data
        in: body
        name: material
        required: true
        schema:
          $ref: '#/definitions/genproto.MaterialCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Material created successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Material
      tags:
      - Material
  /materials/list:
    get:
      consumes:
      - application/json
      description: Retrieve a list of materials
      parameters:
      - description: Lesson ID
        in: query
        name: lesson_id
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of materials
          schema:
            $ref: '#/definitions/genproto.MaterialListRes'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List Materials
      tags:
      - Material
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"